+++ src/core/Breadcrumb/Breadcrumb/Breadcrumb.baseStyles.ts
@@ -1,6 +1,6 @@
 import { css } from 'styled-components';
-import { SuomifiTheme } from '../theme';
-import { nav, list, listItem, font } from '../theme/reset';
+import { SuomifiTheme } from '../../theme';
+import { nav, list, listItem, font } from '../../theme/reset';
 
 export const baseStyles = (theme: SuomifiTheme) => css`
   ${nav(theme)}
