+++ src/core/Breadcrumb/index.ts
@@ -1,0 +1,5 @@
+export { Breadcrumb, BreadcrumbProps } from './Breadcrumb/Breadcrumb';
+export {
+  BreadcrumbLink,
+  BreadcrumbLinkProps,
+} from './BreadcrumbLink/BreadcrumbLink';
